move and tidy files
