+++ backend/src/game-lobby-service/index.ts
@@ -1,0 +1,3 @@
+import GameLobbyService from './GameLobbyService';
+
+export default new GameLobbyService();
